[inductor] ELEMENTS_PER_WARP_32 -> ONE_ELEMENT_PER_THREAD
AMD devices have 64 elements per thread; this PR makes the handling of the "ELEMENTS_PER_WARP_32" generic and uses DeviceProperties.warp_size to determine the warp size instead of hard-coding the warp size as 32. It also renames the enum value. Added a unit test for this.

Note: I left the old enum option (ELEMENTS_PER_WARP_32) as is instead of renaming it. I'm not sure whether we expect should caches to get invalidated here; if this concern is valid, then there's a risk that this would get updated, but some model could use the cached inductor code, which would reference "ELEMENTS_PER_WARP_32", which would no longer exist.

ghstack-source-id: 9c00854afc326788da68e9ff4b853429b429099b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/136472