Fix tensor subclass + dynamic shapes in torch.compile + aot autograd
Fixes issue: #124619

This PR addresses a bug in tensor subclasses and symbolic execution.
For each subclass, it appends the sizes to the list of arguments and
returns the computed shapes at runtime.

Most of the changes are in the `unwrap_tensor_subclasses` function.
This function now takes two extra flags: append_extra and is_runtime. While tracing, if append_extra is true and we are tracing for the forward graph, extra arguments are added.

An extra field (flat_tensor_extra_sizes_offset) is introduced to SubclassCreationMeta. This field stores the offset from right to left for the sizes associated with a tensor subclass. To compute the sizes at runtime, we can use #args[#args - offset : #args - offset + #sizes], where offset is the extra field and #sizes is the number of sizes for the given subclass.

Add tests for two different subclasses: TwoTensor and DoubleTensor. The
latter is a wrapper that behaves as if the inner tensor were twice its
original size.

The set of tests is composed of functions that return a mix of subclasses
and plain tensors.

ghstack-source-id: 97d54cc79c118835281950b01c242aec4d811860
Pull Request resolved: https://github.com/pytorch/pytorch/pull/125941